requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.4.0
djangorestframework-simplejwt==5.3.1
celery==5.4.0
django-celery-beat==2.6.0
drf-yasg==1.21.7
requests==2.32.3
redis==5.0.7
django-cors-headers==4.4.0
django-filter==24.2